requirements.txt
Pyrogram~=2.0.97
pytz~=2022.7.1
python-dateutil~=2.8.2
taskwarrior~=0.1.2
setuptools~=65.5.1